requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.2.1
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.5
reportlab==3.5.59
six==1.15.0
Werkzeug==1.0.1
xlrd==2.0.1
rlextra -i "https://dgbheroku:herokudgb*15@www.reportlab.com/pypi/"
